#1363 move XADataSourceFactory => transaction.xa.manager.dialect
