+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/dialect/XADataSourceFactory.java
@@ -15,11 +15,10 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager;
+package io.shardingsphere.transaction.xa.manager.dialect;
 
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.exception.ShardingException;
-import io.shardingsphere.transaction.xa.manager.dialect.XADatabaseType;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
